Attempt to add more player skins (bugged)
This commit tries to use an MDL file with submodels and separate MDLs.

The separate MDLs are used because they are not included in the main file that contains the submodel.

This commit contains the following errors:

Host_Error: PF_precache_model_I_internal: Bad string '' FATAL ERROR (shutting down): Host_Error: PF_precache_model_I_internal: Bad string ''

I’m making this commit to work on fixing these errors and, if necessary, to be able to revert to this base code.